+++ NEWS.md
@@ -2,7 +2,8 @@
 
 * Log level by default :notice is removed. Use :info. 
 * Added ActArray#power? method
-* Added #beams and #stored attributes for Gripper
+* Added #beams and #stored attributes for Gripper class
+* Added #position, #speed, #acceleration and #current attributes for Actor class
 
 ## 2012-03-06 0.4.1